requirements.txt
python-telegram-bot
requests
numpy